Changing Scope requirements
For construction of the community website bootstrap 4 will be used as it has a cleaner look and is easier to implement CSS3 feature. Designing advance and good-looking CSS code like a Responsive navigation bar is difficult to do from scratch. I will create another Git repo on how to do advance CSS or Advance DOM Manipulation tasks for educational purposes.